--- ios/Flutter/Release.xcconfig
@@ -1,1 +1,0 @@
-#include "Generated.xcconfig"
