--- ios/Runner/GoogleService-Info.plist
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>API_KEY</key>
-	<string>AIzaSyD0p4o729OPNq387xjpE8hg-3wIp9g9nvc</string>
-	<key>GCM_SENDER_ID</key>
-	<string>1006387415222</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.example.weFixIt</string>
-	<key>PROJECT_ID</key>
-	<string>dispositivos-7daf3</string>
-	<key>STORAGE_BUCKET</key>
-	<string>dispositivos-7daf3.appspot.com</string>
-	<key>IS_ADS_ENABLED</key>
-	<false></false>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:1006387415222:ios:8e71879da860db0550be17</string>
-</dict>
-</plist>
