--- ios/Runner/Runner-Bridging-Header.h
@@ -1,1 +1,0 @@
-#import "GeneratedPluginRegistrant.h"
