--- ios/firebase_app_id_file.json
@@ -1,7 +1,0 @@
-{
-  "file_generated_by": "FlutterFire CLI",
-  "purpose": "FirebaseAppID & ProjectID for this Firebase app in this directory",
-  "GOOGLE_APP_ID": "1:1006387415222:ios:8e71879da860db0550be17",
-  "FIREBASE_PROJECT_ID": "dispositivos-7daf3",
-  "GCM_SENDER_ID": "1006387415222"
-}
